refactor: Extract media elements
 - move image from element to media
