+++ shoes-core/lib/shoes/dsl/element.rb
@@ -2,26 +2,6 @@
 class Shoes
   module DSL
     module Element
-      def image(*args, &blk)
-        if blk
-          raise Shoes::NotImplementedError,
-                'Sorry image does not support the block form in Shoes 4!' \
-                ' Check out github issue #1309 for any changes/updates or if you' \
-                ' want to help :)'
-        else
-          opts = style_normalizer.normalize pop_style(args)
-          path, *leftovers = args
-
-          message = <<EOS
-Wrong number of arguments. Must be:
-  - image(path, [opts])
-EOS
-          raise ArgumentError, message if leftovers.any?
-
-          create Shoes::Image, path, opts
-        end
-      end
-
       def border(color, styles = {})
         create Shoes::Border, pattern(color), styles
       end
